Fitness refactor
